// apps/web/src/api/model/libraryResponse.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Scholia API
 * OpenAPI spec version: 0.1.0
 */
import type { LibraryGroup } from "./libraryGroup";
import type { LibraryStats } from "./libraryStats";

export interface LibraryResponse {
    groups: LibraryGroup[];
    stats: LibraryStats;
}